# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DealScanner Control API",
    "description": "Fastify control API for the DealScanner bot and scanner runtime.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/scans/run": {
      "post": {
        "summary": "Run a manual scan pass",
        "tags": [
          "control"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "accepted": {
                      "type": "boolean"
                    },
                    "startedAt": {
                      "type": "string"
                    },
                    "finishedAt": {
                      "type": "string"
                    },
                    "durationMs": {
                      "type": "number"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "manual"
                      ]
                    },
                    "error": {
                      "type": [
                        "null",
                        "string"
                      ]
                    }
                  },
                  "required": [
                    "accepted",
                    "startedAt",
                    "finishedAt",
                    "durationMs",
                    "mode",
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/search-results": {
      "get": {
        "summary": "List recent scanner search results",
        "tags": [
          "control"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "cashConverters",
                "ebay",
                "gumtree",
                "salvos",
                "csMarket",
                "steamMarket",
                "csTradeBot"
              ]
            },
            "in": "query",
            "name": "type",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "queryId",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "properties": {
                          "source": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          },
                          "price": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "imageUrl": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "queryType": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "queryId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "foundAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "source",
                          "title",
                          "url",
                          "price",
                          "imageUrl",
                          "queryType",
                          "queryId",
                          "foundAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "results"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/testing/scans/run": {
      "post": {
        "summary": "Run a temporary scan without persisting a user query",
        "tags": [
          "testing"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "cashConverters",
                      "ebay",
                      "gumtree",
                      "salvos",
                      "csMarket",
                      "steamMarket",
                      "csTradeBot"
                    ]
                  },
                  "payload": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "notify": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "type",
                  "payload"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "startedAt": {
                      "type": "string"
                    },
                    "finishedAt": {
                      "type": "string"
                    },
                    "durationMs": {
                      "type": "number"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "properties": {
                          "source": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          },
                          "price": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "imageUrl": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "passedFilters": {
                            "type": "boolean"
                          },
                          "filterReason": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "source",
                          "title",
                          "url",
                          "price",
                          "imageUrl",
                          "passedFilters",
                          "filterReason"
                        ]
                      }
                    },
                    "notifications": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "properties": {
                          "source": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "imageUrl": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "source",
                          "title",
                          "url"
                        ]
                      }
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "notificationsPublished": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "startedAt",
                    "finishedAt",
                    "durationMs",
                    "items",
                    "notifications",
                    "errors",
                    "notificationsPublished"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}